requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.36
alembic==1.13.2

# Supabase
supabase==2.8.1
postgrest==0.16.0

# File Processing
Pillow==10.4.0
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
xlrd==2.0.1
xlsxwriter==3.2.0

# PDF Processing
pdf2image==1.17.0
pdfkit==1.0.0
weasyprint==61.2
reportlab==4.2.5
PyPDF2==3.0.1
pdfplumber==0.11.0
pymupdf==1.24.12  # PyMuPDF (fitz)

# Image Processing
opencv-python==4.10.0.84
scikit-image==0.24.0
imageio==2.35.2
wand==0.6.14  # ImageMagick wrapper

# Video/Audio Processing
ffmpeg-python==0.2.0
moviepy==1.0.3
pydub==0.25.1

# AI/ML
openai==1.54.3
anthropic==0.39.0
groq==0.11.0
huggingface-hub==0.26.0
transformers==4.44.2
torch==2.5.1
torchvision==0.20.1
tesseract-ocr==0.1.4
pytesseract==0.3.13

# OCR
easyocr==1.7.1

# Document Conversion
mammoth==1.7.0
markdown==3.7
python-pptx==0.6.23
python-rtf==0.45
ebooklib==0.18
pyepub==0.3.7

# Archive Processing
py7zr==0.21.2
rarfile==4.1
patoolib==1.12.0

# Font Processing
fonttools==4.53.1
brotli==1.1.0

# Cloud Services
cloudinary==1.41.4
boto3==1.35.36

# Payment
stripe==11.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
authlib==1.3.2

# HTTP & Requests
httpx==0.27.2
aiohttp==3.10.10
requests==2.32.3

# Utilities
python-dateutil==2.9.0
pytz==2024.2
uuid==1.30

# Logging & Monitoring
structlog==24.4.0
python-json-logger==2.0.7
prometheus-client==0.21.0

# Validation & Error Handling
marshmallow==3.23.1
cerberus==1.3.5

# Environment
python-decouple==3.8

# CORS & Middleware
fastapi-cors==0.1.0

# Task Queue (optional, for heavy processing)
celery==5.4.0
redis==5.2.0

# Compression
gzip==3.12
lzma==3.12

